Fix API call
